Fixed namespace and dependency errors

+++ BirdIdentifier/Controllers/ImageUploadController.cs
@@ -1,7 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using UploadImage_Api.Util;
 
-namespace UploadImage_Api.Controllers;
+using BirdIdentifier.Utils;
+
+namespace BirdIdentifier.Controllers;
 
 [ApiController]
 [Route("images")]
@@ -20,12 +21,17 @@ public class ImageUploadController : ControllerBase
             //Create a checksum used as the file name to avoid duplicate files
             var checksum = ChecksumUtils.GetChecksum(file);
 
+            //Create directory if it doesn't exist
+            Directory.CreateDirectory(@"./UploadedImages");
             var filePath = $@"./UploadedImages/{checksum}{fileExt}";
-
-            //save file to the UploadedImages folder
-            await using (var stream = System.IO.File.Create(filePath))
+            
+            // Save the file if it isn't already present
+            if (!System.IO.File.Exists(filePath))
             {
-                await file.CopyToAsync(stream);
+                await using (var stream = System.IO.File.Create(filePath))
+                {
+                    await file.CopyToAsync(stream);
+                }
             }
 
             //Convert image into format the ML model can use
@@ -33,14 +39,14 @@ public class ImageUploadController : ControllerBase
             {
                 ImageSource = filePath
             };
-
+            
             //Pass in the data and get a prediction
             var predictionResult = MLModel.Predict(mlData);
 
-            //Return prediction to user
+                //Return prediction to user
             return Ok(predictionResult.Prediction);
         }
-
-        return NotFound();
+        
+        return BadRequest("File was not of type .png, .jpg, or .jpeg.");
     }
 }
